Fix(plotting): Bug fixes for issues #587 and #588
* updates to PlotMapView.contour_array() and _VertexCrossSection.contour_array() to allow nan to be passed in the array
* added test_tricountour_NaN() to t007_test.py
* updates to PlotMapView.contour_array() and plot_array() to drop extra dimension from head array with using vertex model grid #588
* updates to UnstructuredPlotUtilities, added irregular_shape_patch() as patch for cross section drawing of vertex and unstructured modelgrids that include multiple geometries (found during fix #588)
* update to geometry.transform() to set x and y dtypes to float (found during fix #588)

+++ flopy/utils/geometry.py
@@ -360,9 +360,9 @@ def transform(x, y, xoff, yoff, angrot_radians,
 
     """
     if isinstance(x, list):
-        x = np.array(x)
+        x = np.array(x, dtype=float)
     if isinstance(y, list):
-        y = np.array(y)
+        y = np.array(y, dtype=float)
 
     if not np.isscalar(x):
         x, y = x.copy(), y.copy()
